Worked on GestureFactory today; it is now able to load dynamic gestures in Windows.  I am using scons (http://www.scons.org/) to build this software, to build under windows you will need to install and properly configure the following:  1) Visual Studio  2) Python  3) SCons
Those are the only dependencies I plan for this software to build it in Windows.  It is possible that I will use cmake in the future; but for now Scons is easier because it directly builds my software; I don't have to generate the proper project files then build from within Visual Studio.